requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
openai>=2.0.0,<3.0.0
python-dotenv>=1.0,<2.0
numpy>=2.0,<3.0
pydantic>=2.0,<3.0
